+++ CHANGELOG.md
@@ -5,6 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 #### Changed
 
+## [0.2.0](https://github.com/KyleGoslan/KGLocationManager/releases/tag/0.2.0)
+- Renamed to 'firstUpdate' to 'isFirstUpdate'.
+- Renamed to 'sharedManager' to 'shared'.
+- Optionally tell the locationManager to start startUpdatingLocation() when calling the configure method (defaults to true). 
+- Can optionally subscribe to two notifications to receive updates (or use delegate methods).
+Released on 2016-22-09.
+
 ## [0.1.2](https://github.com/KyleGoslan/KGLocationManager/releases/tag/0.1.2)
 Released on 2016-22-09.